Calculate a 100 meter extent buffer to query points and lines.
Points don't intersect points or lines so we calculate the extent and use this to intersect feature layers that
are either points or lines.

Selections are naive. We grab the first returned feature without calculating distance to the input point.
Ideally, the code would calculate the distance from the input point to the returned features, but this is beyond scope of demo.